integrate liblfi to use large sandbox verifier

+++ lfi-verify/src/include/lfiv.h
@@ -14,7 +14,7 @@ struct LFIVOptions {
     enum LFIBoxType box;
 
     // Guard size (only used for variable-length sandbox)
-    int32_t guardsize;
+    int64_t guardsize;
 
     // Callback to print a null-terminated error message if verification fails.
     void (*err)(char *msg, size_t size);
